+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## v0.1.0 (2021-04-08)
 * Ported initial codebase from Atlas
+* Split catalogue source to separate interface
+* Split Apache catalogue generator to standalone adapter
+* Added Veneer frontage